+++ src/common/SystemGlobals.h
@@ -67,8 +67,8 @@ extern u8 *pix;
 // Needs to be u32 to accomodation capture and speed
 // buttons
 extern u32 currentButtons[4];
-extern u16 lastButtons[4];
-extern u16 nextButtons[4];
+extern u32 lastButtons[4];
+extern u32 nextButtons[4];
 
 extern int32 sensorX, sensorY;
